[frontend] fix missing deployment count from details (#7328)
