+++ opencti-platform/opencti-front/src/private/components/data/IngestionCatalog/IngestionCatalogConnector.tsx
@@ -9,6 +9,8 @@ import NoConnectorManagersBanner from '@components/data/connectors/NoConnectorMa
 import { Stack } from '@mui/material';
 import IngestionCatalogConnectorCreation from '@components/data/IngestionCatalog/IngestionCatalogConnectorCreation';
 import { IngestionConnector } from '@components/data/IngestionCatalog';
+import createDeploymentCountMap from '@components/data/IngestionCatalog/utils/createDeploymentCountMap';
+import useConnectorDeployDialog from '@components/data/IngestionCatalog/hooks/useConnectorDeployDialog';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
 import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
@@ -24,12 +26,15 @@ const ingestionCatalogConnectorQuery = graphql`
       catalog_id
       contract
     }
+    connectors {
+      manager_contract_image
+    }
   }
 `;
 
 interface IngestionCatalogConnectorComponentProps {
   queryRef: PreloadedQuery<IngestionCatalogConnectorQuery>;
-  onClickDeploy: (connector: IngestionConnector, catalogId: string, hasRegisteredManagers: boolean) => void;
+  onClickDeploy: (connector: IngestionConnector, catalogId: string, hasRegisteredManagers: boolean, deploymentCount: number) => void;
 }
 
 const IngestionCatalogConnectorComponent = ({
@@ -42,7 +47,7 @@ const IngestionCatalogConnectorComponent = ({
 
   const { hasRegisteredManagers } = useConnectorManagerStatus();
 
-  const { contract } = usePreloadedQuery(
+  const { contract, connectors } = usePreloadedQuery(
     ingestionCatalogConnectorQuery,
     queryRef,
   );
@@ -52,6 +57,9 @@ const IngestionCatalogConnectorComponent = ({
   if (!contract) return <ErrorNotFound />;
 
   const connector = JSON.parse(contract.contract);
+
+  const deploymentCounts = createDeploymentCountMap(connectors);
+  const deploymentCount = deploymentCounts.get(connector.container_image) ?? 0;
 
   return (
     <>
@@ -70,7 +78,7 @@ const IngestionCatalogConnectorComponent = ({
         <IngestionCatalogConnectorHeader
           connector={connector}
           isEnterpriseEdition={isEnterpriseEdition}
-          onClickDeploy={() => onClickDeploy(connector, contract?.catalog_id, hasRegisteredManagers)}
+          onClickDeploy={() => onClickDeploy(connector, contract?.catalog_id, hasRegisteredManagers, deploymentCount)}
         />
 
         <IngestionCatalogConnectorOverview connector={connector} />
@@ -79,15 +87,7 @@ const IngestionCatalogConnectorComponent = ({
   );
 };
 
-interface CatalogState {
-  selectedConnector: IngestionConnector | null;
-  selectedCatalogId: string;
-  hasRegisteredManagers: boolean;
-}
-
 const IngestionCatalogConnector = () => {
-  const navigate = useNavigate();
-
   const { connectorSlug } = useParams();
 
   const queryRef = useQueryLoading<IngestionCatalogConnectorQuery>(
@@ -95,28 +95,7 @@ const IngestionCatalogConnector = () => {
     { slug: connectorSlug ?? '' },
   );
 
-  const [catalogState, setCatalogState] = useState<CatalogState>({
-    selectedConnector: null,
-    selectedCatalogId: '',
-    hasRegisteredManagers: false,
-  });
-
-  const handleOpenDeployDialog = (connector: IngestionConnector, catalogId: string, registeredManagers: boolean) => {
-    setCatalogState((prev) => ({
-      ...prev,
-      selectedConnector: connector,
-      selectedCatalogId: catalogId,
-      hasRegisteredManagers: registeredManagers,
-    }));
-  };
-
-  const handleCloseDeployDialog = () => {
-    setCatalogState((prev) => ({
-      ...prev,
-      selectedConnector: null,
-      selectedCatalogId: '',
-    }));
-  };
+  const { catalogState, handleOpenDeployDialog, handleCloseDeployDialog, handleCreate } = useConnectorDeployDialog();
 
   return (
     <Suspense fallback={<Loader variant={LoaderVariant.container} />}>
@@ -137,9 +116,8 @@ const IngestionCatalogConnector = () => {
             onClose={handleCloseDeployDialog}
             catalogId={catalogState.selectedCatalogId}
             hasRegisteredManagers={catalogState.hasRegisteredManagers}
-            onCreate={(connectorId) => {
-              navigate(`${resolveLink('Connectors')}/${connectorId}`);
-            }}
+            onCreate={handleCreate}
+            deploymentCount={catalogState.deploymentCount}
           />
         )
       }
